+++ package.json
@@ -29,11 +29,13 @@
         "@babel/preset-env": "7.14.4",
         "@nuxt/postcss8": "1.1.3",
         "@nuxt/types": "2.15.6",
-        "@nuxt/typescript-build": "2.1.0",
         "@nuxtjs/sitemap": "2.4.0",
         "@types/marked": "4.0.7",
         "autoprefixer": "10.4.12",
         "marked": "4.1.0",
-        "postcss": "8.4.16"
+        "postcss": "8.4.16",
+        "sass": "1.34.0",
+        "sass-loader": "10.2.0",
+        "terser-webpack-plugin": "5.3.6"
     }
 }
